+++ Modules/Intent.Modules.Common.CSharp/release-notes.md
@@ -1,4 +1,8 @@
-﻿### Version 3.8.14
+﻿### Version 3.8.15
+
+- Fixed: `CSharpConstructorCall` incorrectly making things like `null`, `true`, `false`, verbatim identifiers (`@null`).
+
+### Version 3.8.14
 
 - Improvement: Added `Public()` to CSharpClassMethod.
 - Fixed: `CSharpTemplateBase.UseType(...)` would not add usings for generic type arguments.